+++ demo/Spectrogram.js
@@ -17,14 +17,14 @@ class Spectrogram {
         }).catch(e=>console.error('Error getting microphone:', e));
     }
 
-    frame(glsl) {
+    frame(glsl, {viewProjMatrix}) {
         if (!this.analyser) return;
         this.analyser.getByteFrequencyData(this.frequencyArray);
         const n = this.frequencyArray.length;
         const spectro = glsl({size:[n, 1], format:'r8', data:this.frequencyArray, tag:'spectro'});
         const history = glsl({spectro}, 'I.y>0 ? Src(I-ivec2(0,1)) : spectro(ivec2(I.x,0))',
                              {size:[n,256], story:2, wrap:'edge'});
-        glsl({history:history[0], Mesh:history[0].size, DepthTest:1, Perspective:0.5, Aspect:'mean'}, `
+        glsl({viewProjMatrix, history:history[0], Mesh:history[0].size, DepthTest:1, Aspect:'mean'}, `
         varying float z;
         //VERT
         vec4 vertex() {
@@ -32,7 +32,7 @@ class Spectrogram {
             float x = 1.0-log(0.005+UV.x)/log(0.005);
             vec4 pos = vec4((x-0.5)*1.8, UV.y*3.0-0.5, (z-0.5)*0.5, 1.0);
             pos.yz *= rot2(PI/3.0);
-            return pos;
+            return viewProjMatrix*pos;
         }
         //FRAG
         void fragment() {